# contracts/care-coordination-platform-openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Patient Engagement Platform API
  version: 1.0.0
  description: API-first contract examples for the care coordination platform.
servers:
  - url: https://api.healthcare.example.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TelemetryRequest:
      type: object
      required: [deviceId, metricType, metricValue, recordedAt]
      properties:
        deviceId: { type: string }
        metricType: { type: string }
        metricValue: { type: string }
        recordedAt: { type: string, format: date-time }
      example: { deviceId: "dev-22", metricType: "HEART_RATE", metricValue: "124", recordedAt: "2026-05-31T10:05:00Z" }
...
    TelemetryResponse:
      type: object
      required: [id, status, deviceId, metricType, metricValue, recordedAt]
      properties:
        id: { type: string }
        status: { type: string }
        deviceId: { type: string }
        metricType: { type: string }
        metricValue: { type: string }
        recordedAt: { type: string }
      example: { id: "tel-801", status: "RECORDED", deviceId: "dev-22", metricType: "HEART_RATE", metricValue: "124", recordedAt: "2026-05-31T10:05:00Z" }
...
    TelemetryEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/TelemetryResponse' }
    TelemetryListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/TelemetryResponse' }
    TelemetryMetricTypeListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { type: string }
...
    ApiError:
      type: object
      required: [correlationId, code, message, timestamp]
      properties:
        correlationId: { type: string }
        code: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }
      example: { correlationId: "corr-123", code: "RESOURCE_NOT_FOUND", message: "Patient record not found: pat-404", timestamp: "2026-05-31T10:10:00Z" }
  responses:
    BadRequestResponse:
      description: Validation or request error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "VALIDATION_ERROR", message: "Validation failed", timestamp: "2026-05-31T10:11:00Z" }
    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "RESOURCE_NOT_FOUND", message: "Resource not found", timestamp: "2026-05-31T10:11:20Z" }
...
    InternalServerErrorResponse:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "INTERNAL_ERROR", message: "Unexpected server error", timestamp: "2026-05-31T10:11:40Z" }
security:
  - bearerAuth: []
paths:
...
  /telemetry/by-patient/{patientId}:
    get:
      summary: List telemetry readings for a patient
      parameters:
        - in: path
          name: patientId
          required: true
          schema: { type: string }
        - in: query
          name: metricType
          required: false
          schema: { type: string }
        - in: query
          name: startTime
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: endTime
          required: false
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Patient telemetry collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TelemetryListEnvelope' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /telemetry/metric-types:
    get:
      summary: List available telemetry metric types
      parameters:
        - in: query
          name: patientId
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Telemetry metric type collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TelemetryMetricTypeListEnvelope' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /telemetry:
    post:
      summary: Record telemetry reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryRequest'
            examples:
              recordTelemetry:
                value:
                  deviceId: dev-22
                  metricType: HEART_RATE
                  metricValue: '124'
                  recordedAt: 2026-05-31T10:05:00Z
      responses:
        '201':
          description: Telemetry recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TelemetryEnvelope' }
              examples:
                created:
                  value:
                    correlationId: corr-123
                    data:
                      id: tel-801
                      status: RECORDED
                      deviceId: dev-22
                      metricType: HEART_RATE
                      metricValue: '124'
                      recordedAt: 2026-05-31T10:05:00Z
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: List telemetry readings
      responses:
        '200':
          description: Telemetry collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TelemetryListEnvelope' }
              examples:
                list:
                  value:
                    correlationId: corr-123
                    data:
                      - id: tel-801
                        status: RECORDED
                        deviceId: dev-22
                        metricType: HEART_RATE
                        metricValue: '124'
                        recordedAt: 2026-05-31T10:05:00Z
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /telemetry/{id}:
    get:
      summary: Retrieve telemetry reading
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Telemetry retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TelemetryEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: tel-801
                      status: RECORDED
                      deviceId: dev-22
                      metricType: HEART_RATE
                      metricValue: '124'
                      recordedAt: 2026-05-31T10:05:00Z
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
...